Revert "Fix/settings page UI changes"
