+++ webapp/javascript/components/Settings/APIKeys/index.tsx
@@ -40,7 +40,7 @@ const ApiKeys = () => {
           type="submit"
           kind="secondary"
           icon={faPlus}
-          onClick={() => history.push('/settings/api-keys/add')}
+          onClick={() => history.push('/settings/api-key-add')}
         >
           Add Key
         </Button>
@@ -70,7 +70,7 @@ const ApiKeys = () => {
                 <td align="center">
                   <Button
                     type="submit"
-                    kind="danger"
+                    kind="default"
                     aria-label="Delete key"
                     onClick={() => handleDelete(key)}
                   >
