+++ webapp/javascript/components/Settings/Preferences.module.css
@@ -1,3 +1,10 @@
 .preferencesInputWrapper {
   margin-bottom: 30px;
 }
+
+.errors {
+  color: white;
+  background-color: #8a1722;
+  font-weight: bold;
+  padding: 0px 1em 0px 1em;
+}
