+++ webapp/javascript/components/Settings/Preferences.tsx
@@ -1,9 +1,97 @@
 /* eslint-disable prettier/prettier */
-import React from 'react';
+import React, { useState } from 'react';
 import Button from '@ui/Button';
+import { connect } from 'react-redux';
 
-import { withCurrentUser } from '@pyroscope/redux/reducers/user';
+import { useAppDispatch } from '@pyroscope/redux/hooks';
+
+import {
+  changeMyPassword,
+  withCurrentUser,
+} from '@pyroscope/redux/reducers/user';
 import styles from './Preferences.module.css';
+
+function ChangePasswordForm(props) {
+  const { user } = props;
+  const [form, setForm] = useState({ errors: [] });
+  const dispatch = useAppDispatch();
+  if (user.isExternal) {
+    return null;
+  }
+
+  const handleChange = (e) => {
+    setForm({ ...form, [e.target.name]: e.target.value });
+  };
+
+  const onChangePassword = (form) => {
+    dispatch(
+      changeMyPassword({
+        oldPassword: form.oldPassword,
+        newPassword: form.password,
+      })
+    )
+      .unwrap()
+      .then(
+        () => {
+          setForm({
+            errors: [],
+            oldPassword: '',
+            password: '',
+            passwordAgain: '',
+          });
+        },
+        (e) => setForm({ errors: e.errors })
+      );
+  };
+
+  return (
+    <>
+      <h3>Change password</h3>
+      <div>
+        <div className={styles.preferencesInputWrapper}>
+          <div className={styles.errors}>{form.errors.join(', ')}</div>
+          <h4>Old password</h4>
+          <input
+            type="password"
+            placeholder="Password"
+            defaultValue=""
+            name="oldPassword"
+            required
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.preferencesInputWrapper}>
+          <h4>New password</h4>
+          <input
+            type="password"
+            placeholder="Password"
+            defaultValue=""
+            name="password"
+            required
+            onChange={handleChange}
+          />
+        </div>
+        <div className={styles.preferencesInputWrapper}>
+          <h4>New password Again</h4>
+          <input
+            type="password"
+            placeholder="Password"
+            name="passwordAgain"
+            required
+            onChange={handleChange}
+          />
+        </div>
+        <Button
+          type="submit"
+          kind="secondary"
+          onClick={() => onChangePassword(form)}
+        >
+          Save
+        </Button>
+      </div>
+    </>
+  );
+}
 
 function Preferences(props) {
   const { currentUser } = props;
@@ -52,6 +140,8 @@ function Preferences(props) {
       <Button type="submit" kind="secondary">
         Save
       </Button>
+
+      <ChangePasswordForm user={currentUser} />
     </>
   );
 }
