--- webapp/javascript/components/Settings/Security/Security.module.css
@@ -1,10 +1,0 @@
-.securityInputWrapper {
-  margin-bottom: 30px;
-}
-
-.errors {
-  color: white;
-  background-color: #8a1722;
-  font-weight: bold;
-  padding: 0px 1em 0px 1em;
-}
