+++ webapp/javascript/components/Settings/SettingsTable.module.css
@@ -7,14 +7,9 @@
   padding: 15px 0;
 }
 
-.settingsTable th {
-  text-align: left;
-  padding: 10px 20px;
-}
-
 .settingsTable td {
   min-width: 50px;
-  padding: 10px 20px;
+  padding: 10px 0;
 }
 
 .settingsTable tbody {
